requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.3.1
billiard==3.6.3.0
celery==4.3.0
certifi==2020.11.8
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.3
django-admin-honeypot==1.1.0
django-admin-interface==0.14.1
django-colorfield==0.3.2
django-debug-toolbar==3.1.1
django-defender==0.8.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.4.0
django-lockdown==3.0.0
django-loginas==0.3.9
django-multi-captcha-admin==1.0.0
django-recaptcha2==1.4.1
django-request==1.5.6
django-username-email==2.4.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.4.0
jdcal==1.4.1
kombu==4.6.11
MarkupPy==1.14
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.1.5
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
redis==3.2.0
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
django-filter==2.4.0
django-cors-headers==3.7.0
